# Fix cgo argument check panic for small slices from bytes.Buffer (and similar types)