--- a/lmdb/cursor.go
+++ b/lmdb/cursor.go
@@ -210,22 +210,16 @@ func (c *Cursor) Put(key, val []byte, flags uint) error {
 	if len(key) == 0 {
 		return c.putNilKey(flags)
 	}
-	var ret C.int
-	if len(val) != 0 {
-		ret = C.lmdbgo_mdb_cursor_put2(
-			c._c,
-			unsafe.Pointer(&key[0]), C.size_t(len(key)),
-			unsafe.Pointer(&val[0]), C.size_t(len(val)),
-			C.uint(flags),
-		)
-	} else {
-		ret = C.lmdbgo_mdb_cursor_put2(
-			c._c,
-			unsafe.Pointer(&key[0]), C.size_t(len(key)),
-			nil, 0,
-			C.uint(flags),
-		)
+	vn := len(val)
+	if vn == 0 {
+		val = []byte{0}
 	}
+	ret := C.lmdbgo_mdb_cursor_put2(
+		c._c,
+		unsafe.Pointer(&key[0]), C.size_t(len(key)),
+		unsafe.Pointer(&val[0]), C.size_t(len(val)),
+		C.uint(flags),
+	)
 	return operrno("mdb_cursor_put", ret)
 }
 
@@ -261,13 +255,7 @@ func (c *Cursor) PutMulti(key []byte, page []byte, stride int, flags uint) error
 		return c.putNilKey(flags)
 	}
 	if len(page) == 0 {
-		ret := C.lmdbgo_mdb_cursor_putmulti(
-			c._c,
-			unsafe.Pointer(&key[0]), C.size_t(len(key)),
-			nil, 0, C.size_t(stride),
-			C.uint(flags|C.MDB_MULTIPLE),
-		)
-		return operrno("mdb_cursor_put", ret)
+		page = []byte{0}
 	}
 
 	vn := WrapMulti(page, stride).Len()