@bmatsuo 2016-03-23
It seems what performance loses remain here will not be regained until (possibly) go1.7. That is unfortunate. But if nothing else maybe SSA can smooth it all over.

But, I think I need to just bite the bullet and merge this now =\ The possibility of runtime panic is too real and hard to predict/handle.