@mranney 2016-03-23
That's annoying to lose some perf, but I support your decision to favor stability. :)