@bmatsuo 2016-03-23
Indeed. At least I don't think the slowdown is a killer -- 100ns per OP. It should be OK for the time being. LMDB should still handily outperform BoltDB with these changes that gulf in performance is pretty large.